move some types around
