+++ src/configuration/types.ts
@@ -1,5 +1,5 @@
-import { IFormatterConfiguration } from '../formatter'
 import { IRuntimeOptions } from '../runtime'
+import { IParsedArgvFormatOptions } from '../cli/argv_parser'
 
 export interface IRunConfiguration {
   sources: {
@@ -17,4 +17,16 @@ export interface IRunConfiguration {
   }
   runtime: IRuntimeOptions & { parallel: number }
   formats?: IFormatterConfiguration
+}
+
+export interface IFormatterConfiguration {
+  stdout?: string
+  files?: Record<string, string>
+  publish?:
+    | {
+        url?: string
+        token?: string
+      }
+    | false
+  options?: IParsedArgvFormatOptions
 }
